requirements.txt
sympy
Pycryptodome
pwntools
gmpy2
sage
